Some refactoring

+++ VideoConverter/Converter/Converter.cs
@@ -1,5 +1,6 @@
 using System;
 using System.IO;
+using System.Linq;
 using System.Threading.Tasks;
 using Xabe.FFmpeg;
 using Xabe.FFmpeg.Exceptions;
@@ -8,20 +9,18 @@ namespace VideoConverter;
 
 public static class Converter
 {
-    private static readonly string _alternativeFFmpegPath = Environment.CurrentDirectory;
+    private static readonly string[] _executables = new string[] { "ffmpeg.exe", "ffprobe.exe", "ffplay.exe" };
+    private static readonly string _executablesPath = Environment.CurrentDirectory;
 
-    public static async Task<string> ConvertVideoAsync(string inputFilePath, string outputFileDir, string outputFormat)
+    public static async Task<string> ConvertAsync(string inputFilePath, string outputFileDir, string outputFormat)
     {
-        if (Utility.FFmpegExecutablesExist(_alternativeFFmpegPath))
+        if (FFmpegExecutablesExist(_executablesPath))
         {
-            FFmpeg.SetExecutablesPath(_alternativeFFmpegPath);
+            FFmpeg.SetExecutablesPath(_executablesPath);
         }
 
         var outputFilePath = Utility.GetOutputFilepath(inputFilePath, outputFileDir, outputFormat);
-        if (File.Exists(outputFilePath))
-        {
-            File.Delete(outputFilePath);
-        }
+        if (File.Exists(outputFilePath)) File.Delete(outputFilePath);
 
         try
         {
@@ -46,11 +45,17 @@ public static class Converter
         }
         catch (FFmpegNotFoundException)
         {
-            throw new FFmpegPathException($"FFmpeg executables not found in environment PATH or in {_alternativeFFmpegPath}.");
+            throw new FFmpegPathException($"FFmpeg executables not found in environment PATH or in {_executablesPath}.");
         }
         catch (Exception ex)
         {
             throw new ConversionException("Conversion failed.", ex);
         }
+    }
+
+    private static bool FFmpegExecutablesExist(string targetDirectory)
+    {
+        var files = Directory.GetFiles(targetDirectory).Select(Path.GetFileName);
+        return _executables.All(x => files.Contains(x));
     }
 }
